Created Provider for Streams
